requirements.txt
# Slack integration
slack_bolt>=1.23.0
slack_sdk>=3.35.0

# Utilities
python-dotenv>=1.1.0
httpx>=0.24.1
aiohttp>=3.11.0
uvicorn>=0.23.2

# MCP Related
mcp>=1.0.0
mcp-agent>=0.0.16  # Latest version

# Development dependencies
pyright>=1.1.379
pytest>=8.3.0
ruff>=0.6.0
